Refactor warlock application to support multiple similarity algorithms and enhance argument parsing
